@@ -1,13 +1,16 @@
-// swift-tools-version: 5.7
-// The swift-tools-version declares the minimum version of Swift required to build this package.
+// swift-tools-version: 6.0
 
 import PackageDescription
 
+#if os(macOS)
+let platforms: [PackageDescription.SupportedPlatform]? = [.macOS(.v15)]
+#else
+let platforms: [PackageDescription.SupportedPlatform]? = nil
+#endif
+
 let package = Package(
     name: "BreezeLambdaDynamoDBAPI",
-    platforms: [
-        .macOS(.v13),
-    ],
+    platforms: platforms,
     products: [
         .library(
             name: "BreezeDynamoDBService",
