@@ -1,4 +1,4 @@
-//    Copyright 2023 (c) Andrea Scuderi - https://github.com/swift-serverless
+//    Copyright 2024 (c) Andrea Scuderi - https://github.com/swift-serverless
 //
 //    Licensed under the Apache License, Version 2.0 (the "License");
 //    you may not use this file except in compliance with the License.
@@ -17,18 +17,34 @@ import AWSLambdaRuntime
 import BreezeDynamoDBService
 import Logging
 
-struct BreezeLambdaHandler<T: BreezeCodable> {
+/// BreezeLambdaHandler implements a Lambda handler for Breeze operations.
+/// It conforms to the `LambdaHandler` protocol and is generic over a type `T` that conforms to `BreezeCodable`.
+///
+/// This handler supports the following operations:
+///
+/// - Create: Creates a new item in the DynamoDB table.
+/// - Read: Reads an item from the DynamoDB table based on the provided key.
+/// - Update: Updates an existing item in the DynamoDB table.
+/// - Delete: Deletes an item from the DynamoDB table based on the provided key and timestamps.
+/// - List: Lists items in the DynamoDB table with optional pagination.
+struct BreezeLambdaHandler<T: BreezeCodable>: LambdaHandler, Sendable {
     typealias Event = APIGatewayV2Request
     typealias Output = APIGatewayV2Response
 
-    let service: BreezeDynamoDBServing
+    let dbManager: BreezeDynamoDBManaging
     let operation: BreezeOperation
 
     var keyName: String {
-        self.service.keyName
+        self.dbManager.keyName
     }
-
-    func handle(context: AWSLambdaRuntimeCore.LambdaContext, event: APIGatewayV2Request) async -> APIGatewayV2Response {
+    
+    /// Lambda handler for Breeze operations.
+    /// - Parameters:
+    ///  - event: The event containing the API Gateway request.
+    ///  - context: The Lambda context providing information about the invocation.
+    ///
+    /// This initializer sets up the Breeze Lambda handler with the specified DynamoDB manager and operation.
+    func handle(_ event: APIGatewayV2Request, context: LambdaContext) async throws -> APIGatewayV2Response {
         switch self.operation {
         case .create:
             return await self.createLambdaHandler(context: context, event: event)
@@ -43,39 +59,42 @@ struct BreezeLambdaHandler<T: BreezeCodable> {
         }
     }
 
-    func createLambdaHandler(context: AWSLambdaRuntimeCore.LambdaContext, event: APIGatewayV2Request) async -> APIGatewayV2Response {
+    /// Lambda handler for creating an item in the DynamoDB table.
+    func createLambdaHandler(context: LambdaContext, event: APIGatewayV2Request) async -> APIGatewayV2Response {
         guard let item: T = try? event.bodyObject() else {
             let error = BreezeLambdaAPIError.invalidRequest
             return APIGatewayV2Response(with: error, statusCode: .forbidden)
         }
         do {
-            let result: T = try await service.createItem(item: item)
+            let result: T = try await dbManager.createItem(item: item)
             return APIGatewayV2Response(with: result, statusCode: .created)
         } catch {
             return APIGatewayV2Response(with: error, statusCode: .forbidden)
         }
     }
 
-    func readLambdaHandler(context: AWSLambdaRuntimeCore.LambdaContext, event: APIGatewayV2Request) async -> APIGatewayV2Response {
+    /// Lambda handler for reading an item from the DynamoDB table.
+    func readLambdaHandler(context: LambdaContext, event: APIGatewayV2Request) async -> APIGatewayV2Response {
         guard let key = event.pathParameters?[keyName] else {
             let error = BreezeLambdaAPIError.invalidRequest
             return APIGatewayV2Response(with: error, statusCode: .forbidden)
         }
         do {
-            let result: T = try await service.readItem(key: key)
+            let result: T = try await dbManager.readItem(key: key)
             return APIGatewayV2Response(with: result, statusCode: .ok)
         } catch {
             return APIGatewayV2Response(with: error, statusCode: .notFound)
         }
     }
 
-    func updateLambdaHandler(context: AWSLambdaRuntimeCore.LambdaContext, event: APIGatewayV2Request) async -> APIGatewayV2Response {
+    /// Lambda handler for updating an item in the DynamoDB table.
+    func updateLambdaHandler(context: LambdaContext, event: APIGatewayV2Request) async -> APIGatewayV2Response {
         guard let item: T = try? event.bodyObject() else {
             let error = BreezeLambdaAPIError.invalidRequest
             return APIGatewayV2Response(with: error, statusCode: .forbidden)
         }
         do {
-            let result: T = try await service.updateItem(item: item)
+            let result: T = try await dbManager.updateItem(item: item)
             return APIGatewayV2Response(with: result, statusCode: .ok)
         } catch {
             return APIGatewayV2Response(with: error, statusCode: .notFound)
@@ -88,7 +107,8 @@ struct BreezeLambdaHandler<T: BreezeCodable> {
         var updatedAt: String?
     }
 
-    func deleteLambdaHandler(context: AWSLambdaRuntimeCore.LambdaContext, event: APIGatewayV2Request) async -> APIGatewayV2Response {
+    /// Lambda handler for deleting an item from the DynamoDB table.
+    func deleteLambdaHandler(context: LambdaContext, event: APIGatewayV2Request) async -> APIGatewayV2Response {
         guard let key = event.pathParameters?[keyName],
               let createdAt = event.queryStringParameters?["createdAt"],
               let updatedAt = event.queryStringParameters?["updatedAt"] else {
@@ -97,18 +117,19 @@ struct BreezeLambdaHandler<T: BreezeCodable> {
         }
         do {
             let simpleItem = SimpleItem(key: key, createdAt: createdAt, updatedAt: updatedAt)
-            try await self.service.deleteItem(item: simpleItem)
+            try await self.dbManager.deleteItem(item: simpleItem)
             return APIGatewayV2Response(with: BreezeEmptyResponse(), statusCode: .ok)
         } catch {
             return APIGatewayV2Response(with: error, statusCode: .notFound)
         }
     }
 
-    func listLambdaHandler(context: AWSLambdaRuntimeCore.LambdaContext, event: APIGatewayV2Request) async -> APIGatewayV2Response {
+    /// Lambda handler for listing items in the DynamoDB table.
+    func listLambdaHandler(context: LambdaContext, event: APIGatewayV2Request) async -> APIGatewayV2Response {
         do {
             let key = event.queryStringParameters?["exclusiveStartKey"]
             let limit: Int? = event.queryStringParameter("limit")
-            let result: ListResponse<T> = try await service.listItems(key: key, limit: limit)
+            let result: ListResponse<T> = try await dbManager.listItems(key: key, limit: limit)
             return APIGatewayV2Response(with: result, statusCode: .ok)
         } catch {
             return APIGatewayV2Response(with: error, statusCode: .forbidden)
