@@ -1,4 +1,4 @@
-//    Copyright 2023 (c) Andrea Scuderi - https://github.com/swift-serverless
+//    Copyright 2024 (c) Andrea Scuderi - https://github.com/swift-serverless
 //
 //    Licensed under the Apache License, Version 2.0 (the "License");
 //    you may not use this file except in compliance with the License.
@@ -12,195 +12,61 @@
 //    See the License for the specific language governing permissions and
 //    limitations under the License.
 
-import SotoCore
 import SotoDynamoDB
-import XCTest
+import AsyncHTTPClient
+import Logging
+import Testing
 @testable import BreezeDynamoDBService
 
-struct Product: BreezeCodable {
-    var key: String
-    var name: String
-    var description: String
-    var createdAt: String?
-    var updatedAt: String?
-}
-
-final class BreezeDynamoDBServiceTests: XCTestCase {
-    
-    let tableName = "Breeze"
-    let keyName = "key"
-    var sut: BreezeDynamoDBService!
-    
-    let product2023 = Product(key: "2023", name: "Swift Serverless API 2022", description: "Test")
-    let product2022 = Product(key: "2022", name: "Swift Serverless API with async/await! 🚀🥳", description: "BreezeLambaAPI is magic 🪄!")
-    
-    override func setUp() async throws {
-        try await super.setUp()
-        try await LocalStackDynamoDB.createTable(name: tableName, keyName: keyName)
-        let db = LocalStackDynamoDB.dynamoDB
-        sut = BreezeDynamoDBService(db: db, tableName: tableName, keyName: keyName)
-    }
-
-    override func tearDown() async throws {
-        sut = nil
-        try await LocalStackDynamoDB.deleteTable(name: tableName)
-        try await super.tearDown()
+@Suite
+struct BreezeDynamoDBServiceTests {
+    @Test
+    func testInitPrepareBreezeDynamoDBManager() async throws {
+        let sut = await makeBreezeDynamoDBConfig()
+        let manager = await sut.dbManager()
+        #expect(manager is BreezeDynamoDBManager, "Expected BreezeDynamoDBManager instance")
+        try await sut.gracefulShutdown()
     }
     
-    func test_createItem() async throws {
-        let value = try await sut.createItem(item: product2023)
-        XCTAssertEqual(value.key, product2023.key)
-        XCTAssertEqual(value.name, product2023.name)
-        XCTAssertEqual(value.description, product2023.description)
-        XCTAssertNotNil(value.createdAt?.iso8601)
-        XCTAssertNotNil(value.updatedAt?.iso8601)
+    @Test
+    func testGracefulShutdownCanBeCalledMultipleTimes() async throws {
+        let sut = await makeBreezeDynamoDBConfig()
+        try await sut.gracefulShutdown()
+        try await sut.gracefulShutdown()
     }
     
-    func test_createItemDuplicate_shouldThrowConditionalCheckFailedException() async throws {
-        let value = try await sut.createItem(item: product2023)
-        XCTAssertEqual(value.key, product2023.key)
-        XCTAssertEqual(value.name, product2023.name)
-        XCTAssertEqual(value.description, product2023.description)
-        XCTAssertNotNil(value.createdAt?.iso8601)
-        XCTAssertNotNil(value.updatedAt?.iso8601)
-        do {
-            _ = try await sut.createItem(item: product2023)
-            XCTFail("It should throw conditionalCheckFailedException")
-        } catch {
-            XCTAssertNotNil(error)
-        }
+    @Test
+    func testMockInjection() async throws {
+        let config = BreezeDynamoDBConfig(
+            region: .useast1,
+            tableName: "TestTable",
+            keyName: "TestKey",
+        )
+        let logger = Logger(label: "BreezeDynamoDBServiceTests")
+        let httpConfig = BreezeHTTPClientConfig(timeout: .seconds(10), logger: logger)
+        let sut = await BreezeDynamoDBService(
+            config: config,
+            httpConfig: httpConfig,
+            logger: logger,
+            DBManagingType: BreezeDynamoDBManagerMock.self
+        )
+        let manager = await sut.dbManager()
+        #expect(manager is BreezeDynamoDBManagerMock, "Expected BreezeDynamoDBManager instance")
+        try await sut.gracefulShutdown()
     }
     
-    func test_readItem() async throws {
-        let cretedItem = try await sut.createItem(item: product2023)
-        let readedItem: Product = try await sut.readItem(key: "2023")
-        XCTAssertEqual(cretedItem.key, readedItem.key)
-        XCTAssertEqual(cretedItem.name, readedItem.name)
-        XCTAssertEqual(cretedItem.description, readedItem.description)
-        XCTAssertEqual(cretedItem.createdAt?.iso8601, readedItem.createdAt?.iso8601)
-        XCTAssertEqual(cretedItem.updatedAt?.iso8601, readedItem.updatedAt?.iso8601)
-    }
-    
-    func test_readItem_whenItemIsMissing() async throws {
-        let value = try await sut.createItem(item: product2023)
-        XCTAssertEqual(value.key, "2023")
-        do {
-            let _: Product = try await sut.readItem(key: "2022")
-            XCTFail("It should throw when Item is missing")
-        } catch {
-            XCTAssertNotNil(error)
-        }
-    }
-    
-    func test_updateItem() async throws {
-        var value = try await sut.createItem(item: product2023)
-        value.name = "New Name"
-        value.description = "New Description"
-        let newValue = try await sut.updateItem(item: value)
-        XCTAssertEqual(value.key, newValue.key)
-        XCTAssertEqual(value.name, newValue.name)
-        XCTAssertEqual(value.description, newValue.description)
-        XCTAssertEqual(value.createdAt?.iso8601, newValue.createdAt?.iso8601)
-        XCTAssertNotEqual(value.updatedAt?.iso8601, newValue.updatedAt?.iso8601)
-    }
-    
-    func test_updateItem_whenItemHasChanged_shouldThrowConditionalCheckFailedException() async throws {
-        var value = try await sut.createItem(item: product2023)
-        value.name = "New Name"
-        value.description = "New Description"
-        let newValue = try await sut.updateItem(item: value)
-        XCTAssertEqual(value.key, newValue.key)
-        XCTAssertEqual(value.name, newValue.name)
-        XCTAssertEqual(value.description, newValue.description)
-        XCTAssertEqual(value.createdAt?.iso8601, newValue.createdAt?.iso8601)
-        XCTAssertNotEqual(value.updatedAt?.iso8601, newValue.updatedAt?.iso8601)
-        do {
-            let _: Product = try await sut.updateItem(item: product2023)
-            XCTFail("It should throw conditionalCheckFailedException")
-        } catch {
-            XCTAssertNotNil(error)
-        }
-        
-        do {
-            let _: Product = try await sut.updateItem(item: product2022)
-            XCTFail("It should throw conditionalCheckFailedException")
-        } catch {
-            XCTAssertNotNil(error)
-        }
-    }
-    
-    func test_deleteItem() async throws {
-        let value = try await sut.createItem(item: product2023)
-        XCTAssertEqual(value.key, "2023")
-        try await sut.deleteItem(item: value)
-        let readedItem: Product? = try? await sut.readItem(key: "2023")
-        XCTAssertNil(readedItem)
-    }
-    
-    func test_deleteItem_whenItemIsMissing_thenShouldThrow() async throws {
-        do {
-            try await sut.deleteItem(item: product2022)
-            XCTFail("It should throw ServiceError.missingParameters")
-        } catch {
-            XCTAssertNotNil(error)
-        }
-    }
-    
-    func test_deleteItem_whenMissingUpdatedAt_thenShouldThrow() async throws {
-        var value = try await sut.createItem(item: product2023)
-        XCTAssertEqual(value.key, "2023")
-        value.updatedAt = nil
-        do {
-            try await sut.deleteItem(item: value)
-            XCTFail("It should throw ServiceError.missingParameters")
-        } catch {
-            XCTAssertNotNil(error)
-        }
-    }
-    
-    func test_deleteItem_whenMissingCreatedAt_thenShouldThrow() async throws {
-        var value = try await sut.createItem(item: product2023)
-        XCTAssertEqual(value.key, "2023")
-        value.createdAt = nil
-        do {
-            try await sut.deleteItem(item: value)
-            XCTFail("It should throw ServiceError.missingParameters")
-        } catch {
-            XCTAssertNotNil(error)
-        }
-    }
-    
-    func test_deleteItem_whenOutdatedUpdatedAt_thenShouldThrow() async throws {
-        var value = try await sut.createItem(item: product2023)
-        XCTAssertEqual(value.key, "2023")
-        value.updatedAt = Date().iso8601
-        do {
-            try await sut.deleteItem(item: value)
-            XCTFail("It should throw ServiceError.missingParameters")
-        } catch {
-            XCTAssertNotNil(error)
-        }
-    }
-    
-    func test_deleteItem_whenOutdatedCreatedAt_thenShouldThrow() async throws {
-        var value = try await sut.createItem(item: product2023)
-        XCTAssertEqual(value.key, "2023")
-        value.createdAt = Date().iso8601
-        do {
-            try await sut.deleteItem(item: value)
-            XCTFail("It should throw ServiceError.missingParameters")
-        } catch {
-            XCTAssertNotNil(error)
-        }
-    }
-    
-    func test_listItem() async throws {
-        let value1 = try await sut.createItem(item: product2022)
-        let value2 = try await sut.createItem(item: product2023)
-        let list: ListResponse<Product> = try await sut.listItems(key: nil, limit: nil)
-        XCTAssertEqual(list.items.count, 2)
-        let keys = Set(list.items.map { $0.key })
-        XCTAssertTrue(keys.contains(value1.key))
-        XCTAssertTrue(keys.contains(value2.key))
+    private func makeBreezeDynamoDBConfig() async -> BreezeDynamoDBService {
+        let config = BreezeDynamoDBConfig(
+            region: .useast1,
+            tableName: "TestTable",
+            keyName: "TestKey",
+        )
+        let logger = Logger(label: "BreezeDynamoDBServiceTests")
+        let httpConfig = BreezeHTTPClientConfig(timeout: .seconds(10), logger: logger)
+        return await BreezeDynamoDBService(
+            config: config,
+            httpConfig: httpConfig,
+            logger: logger,
+        )
     }
 }
