@@ -1,4 +1,4 @@
-//    Copyright 2023 (c) Andrea Scuderi - https://github.com/swift-serverless
+//    Copyright 2024 (c) Andrea Scuderi - https://github.com/swift-serverless
 //
 //    Licensed under the Apache License, Version 2.0 (the "License");
 //    you may not use this file except in compliance with the License.
@@ -15,4 +15,5 @@
 import class Foundation.JSONDecoder
 import class Foundation.JSONEncoder
 
+/// A simple struct representing an empty response for Breeze Lambda API.
 struct BreezeEmptyResponse: Codable {}
