@@ -1,4 +1,4 @@
-//    Copyright 2023 (c) Andrea Scuderi - https://github.com/swift-serverless
+//    Copyright 2024 (c) Andrea Scuderi - https://github.com/swift-serverless
 //
 //    Licensed under the Apache License, Version 2.0 (the "License");
 //    you may not use this file except in compliance with the License.
@@ -12,113 +12,69 @@
 //    See the License for the specific language governing permissions and
 //    limitations under the License.
 
-import AsyncHTTPClient
-import AWSLambdaEvents
-import AWSLambdaRuntimeCore
-import BreezeDynamoDBService
-import Foundation
 import SotoDynamoDB
+import ServiceLifecycle
+import BreezeDynamoDBService
+import AWSLambdaRuntime
 
-public extension LambdaInitializationContext {
-    enum DynamoDB {
-        public static var Service: BreezeDynamoDBServing.Type = BreezeDynamoDBService.self
-        public static var dbTimeout: Int64 = 30
-    }
-}
-
-public class BreezeLambdaAPI<T: BreezeCodable>: LambdaHandler {
-    public typealias Event = APIGatewayV2Request
-    public typealias Output = APIGatewayV2Response
-
-    let dbTimeout: Int64
-    let region: Region
-    let db: SotoDynamoDB.DynamoDB
-    let service: BreezeDynamoDBServing
-    let tableName: String
-    let keyName: String
-    let operation: BreezeOperation
-    var httpClient: HTTPClient
-
-    static func currentRegion() -> Region {
-        if let awsRegion = Lambda.env("AWS_REGION") {
-            let value = Region(rawValue: awsRegion)
-            return value
-        } else {
-            return .useast1
+/// BreezeLambdaAPI is a service that integrates with AWS Lambda to provide a Breeze API for DynamoDB operations.
+/// It supports operations such as create, read, update, delete, and list items in a DynamoDB table using a BreezeCodable.
+/// This Service is designed to work with ServiceLifecycle, allowing it to be run and stopped gracefully.
+public actor BreezeLambdaAPI<T: BreezeCodable>: Service {
+    
+    let logger = Logger(label: "service-group-breeze-lambda-api")
+    let timeout: TimeAmount
+    private let serviceGroup: ServiceGroup
+    private let apiConfig: any APIConfiguring
+    
+    /// Initializes the BreezeLambdaAPI with the provided API configuration.
+    /// - Parameter apiConfig: An object conforming to `APIConfiguring` that provides the necessary configuration for the Breeze API.
+    /// - Throws: An error if the configuration is invalid or if the service fails to initialize.
+    ///
+    /// This initializer sets up the Breeze Lambda API service with the specified configuration, including the DynamoDB service and the operation to be performed.
+    ///
+    /// - Note:
+    ///   - The `apiConfig` parameter must conform to the `APIConfiguring` protocol, which provides the necessary configuration details for the Breeze API.
+    ///   - The default implementation uses `BreezeAPIConfiguration`, but you can provide your own implementation if needed.
+    public init(apiConfig: APIConfiguring = BreezeAPIConfiguration()) async throws {
+        do {
+            self.apiConfig = apiConfig
+            self.timeout = .seconds(apiConfig.dbTimeout)
+            let config = try apiConfig.getConfig()
+            let httpConfig = BreezeHTTPClientConfig(
+                timeout: .seconds(60),
+                logger: logger
+            )
+            let operation = try apiConfig.operation()
+            let dynamoDBService = await BreezeDynamoDBService(
+                config: config,
+                httpConfig: httpConfig,
+                logger: logger
+            )
+            let breezeLambdaService = BreezeLambdaService<T>(
+                dynamoDBService: dynamoDBService,
+                operation: operation,
+                logger: logger
+            )
+            self.serviceGroup = ServiceGroup(
+                services: [breezeLambdaService],
+                gracefulShutdownSignals: [.sigterm, .sigint],
+                logger: logger
+            )
+        } catch {
+            logger.error("\(error.localizedDescription)")
+            throw error
         }
     }
-
-    static func tableName() throws -> String {
-        guard let tableName = Lambda.env("DYNAMO_DB_TABLE_NAME") else {
-            throw BreezeLambdaAPIError.tableNameNotFound
-        }
-        return tableName
-    }
-
-    static func keyName() throws -> String {
-        guard let tableName = Lambda.env("DYNAMO_DB_KEY") else {
-            throw BreezeLambdaAPIError.keyNameNotFound
-        }
-        return tableName
-    }
-
-    public required init(context: LambdaInitializationContext) async throws {
-        guard let handler = Lambda.env("_HANDLER"),
-              let operation = BreezeOperation(handler: handler)
-        else {
-            throw BreezeLambdaAPIError.invalidHandler
-        }
-        self.operation = operation
-        context.logger.info("operation: \(operation)")
-        self.region = Self.currentRegion()
-        context.logger.info("region: \(region)")
-        self.dbTimeout = LambdaInitializationContext.DynamoDB.dbTimeout
-        context.logger.info("dbTimeout: \(dbTimeout)")
-        self.tableName = try Self.tableName()
-        context.logger.info("tableName: \(tableName)")
-        self.keyName = try Self.keyName()
-        context.logger.info("keyName: \(keyName)")
-
-        let lambdaRuntimeTimeout: TimeAmount = .seconds(dbTimeout)
-        let timeout = HTTPClient.Configuration.Timeout(
-            connect: lambdaRuntimeTimeout,
-            read: lambdaRuntimeTimeout
-        )
-
-        let configuration = HTTPClient.Configuration(timeout: timeout)
-        self.httpClient = HTTPClient(
-            eventLoopGroupProvider: .shared(context.eventLoop),
-            configuration: configuration
-        )
-
-        let awsClient = AWSClient(httpClientProvider: .shared(self.httpClient))
-        self.db = SotoDynamoDB.DynamoDB(client: awsClient, region: self.region)
-
-        self.service = LambdaInitializationContext.DynamoDB.Service.init(
-            db: self.db,
-            tableName: self.tableName,
-            keyName: self.keyName
-        )
-
-        context.terminator.register(name: "shutdown") { eventLoop in
-            context.logger.info("shutdown: started")
-            let promise = eventLoop.makePromise(of: Void.self)
-            Task {
-                do {
-                    try awsClient.syncShutdown()
-                    try await self.httpClient.shutdown()
-                    promise.succeed()
-                    context.logger.info("shutdown: succeed")
-                } catch {
-                    promise.fail(error)
-                    context.logger.info("shutdown: fail")
-                }
-            }
-            return promise.futureResult
-        }
-    }
-
-    public func handle(_ event: AWSLambdaEvents.APIGatewayV2Request, context: AWSLambdaRuntimeCore.LambdaContext) async throws -> AWSLambdaEvents.APIGatewayV2Response {
-        return await BreezeLambdaHandler<T>(service: self.service, operation: self.operation).handle(context: context, event: event)
+    
+    /// Runs the BreezeLambdaAPI service.
+    /// This method starts the internal ServiceGroup and begins processing requests.
+    /// - Throws: An error if the service fails to start or if an issue occurs during execution.
+    ///
+    /// The internal ServiceGroup will handle the lifecycle of the BreezeLambdaAPI, including starting and stopping the service gracefully.
+    public func run() async throws {
+        logger.info("Starting BreezeLambdaAPI...")
+        try await serviceGroup.run()
+        logger.info("BreezeLambdaAPI is stopped successfully")
     }
 }
