@@ -1,4 +1,4 @@
-//    Copyright 2023 (c) Andrea Scuderi - https://github.com/swift-serverless
+//    Copyright 2024 (c) Andrea Scuderi - https://github.com/swift-serverless
 //
 //    Licensed under the Apache License, Version 2.0 (the "License");
 //    you may not use this file except in compliance with the License.
@@ -16,43 +16,48 @@ import BreezeDynamoDBService
 @testable import BreezeLambdaAPI
 import SotoDynamoDB
 
-struct BreezeDynamoDBServiceMock: BreezeDynamoDBServing {
-    var keyName: String
-
-    static var response: (any BreezeCodable)?
-    static var keyedResponse: (any BreezeCodable)?
-
+actor BreezeDynamoDBManagerMock: BreezeDynamoDBManaging {
+    let keyName: String
+    
+    private var response: (any BreezeCodable)?
+    private var keyedResponse: (any BreezeCodable)?
+    
+    func setupMockResponse(response: (any BreezeCodable)?, keyedResponse: (any BreezeCodable)?) {
+        self.keyedResponse = keyedResponse
+        self.response = response
+    }
+    
     init(db: SotoDynamoDB.DynamoDB, tableName: String, keyName: String) {
         self.keyName = keyName
     }
-
+    
     func createItem<T: BreezeCodable>(item: T) async throws -> T {
-        guard let response = Self.response as? T else {
+        guard let response = self.response as? T else {
             throw BreezeLambdaAPIError.invalidRequest
         }
         return response
     }
-
+    
     func readItem<T: BreezeCodable>(key: String) async throws -> T {
-        guard let response = Self.keyedResponse as? T,
+        guard let response = self.keyedResponse as? T,
               response.key == key
         else {
             throw BreezeLambdaAPIError.invalidRequest
         }
         return response
     }
-
+    
     func updateItem<T: BreezeCodable>(item: T) async throws -> T {
-        guard let response = Self.keyedResponse as? T,
+        guard let response = self.keyedResponse as? T,
               response.key == item.key
         else {
             throw BreezeLambdaAPIError.invalidRequest
         }
         return response
     }
-
+    
     func deleteItem<T: BreezeCodable>(item: T) async throws {
-        guard let response = Self.keyedResponse,
+        guard let response = self.keyedResponse,
               response.key == item.key,
               response.createdAt == item.createdAt,
               response.updatedAt == item.updatedAt
@@ -61,22 +66,15 @@ struct BreezeDynamoDBServiceMock: BreezeDynamoDBServing {
         }
         return
     }
-
-    static var limit: Int?
-    static var exclusiveKey: String?
+    
+    var limit: Int?
+    var exclusiveKey: String?
     func listItems<T: BreezeCodable>(key: String?, limit: Int?) async throws -> ListResponse<T> {
-        guard let response = Self.response as? T else {
+        guard let response = self.response as? T else {
             throw BreezeLambdaAPIError.invalidItem
         }
-        Self.limit = limit
-        Self.exclusiveKey = key
+        self.limit = limit
+        self.exclusiveKey = key
         return ListResponse(items: [response], lastEvaluatedKey: key)
-    }
-
-    static func reset() {
-        Self.limit = nil
-        Self.exclusiveKey = nil
-        Self.response = nil
-        Self.keyedResponse = nil
     }
 }
